requirements.txt
requests
pyyaml==5.4.1
pathlib
pyTelegramBotAPI
rich
joblib